requirements.txt
tqdm
pdf2image
Pillow
requests
ultralytics
